+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Inlay Parameters:
 
 1. Don't display inlay parameter hints when a `UnaryLookup` or `Lookup` name matches the parameter name.
+1. Don't display inlay parameter hints when the name at the end of a `SimplMapExpr` matches the parameter name.
 
 Code Folding: